+++ NEWS.md
@@ -5,6 +5,8 @@ NEWS
 
 - Found ugly bug in emis_grid!
 - Added test for emis_grid
+- Add streets profile in emis_to_streets
+
 
 # vein 0.7.10  (Release date: 2019-06-21)